requirements.txt
asgiref==3.3.2
Django==3.1.7
django-cors-headers==3.7.0
django-phone-field==1.8.1
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
